requirements.txt
aniso8601==3.0.2
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.1
backports.functools-lru-cache==1.5
cffi==1.11.5
Click==7.0
configparser==3.5.0
coverage==4.5.1
DateTime==4.3
enum34==1.1.6
Flask==1.0.2
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
passlib==1.7.1
pathlib2==2.3.2
pkg-resources==0.0.0
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pylint==1.9.3
pytest==3.9.1
pytest-cov==2.6.0
pytz==2018.5
scandir==1.9.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.5.0
